web-app-admin/crawler_admin/requirements.txt
APScheduler==3.7.0
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_05nm_gqf36/croots/recipe/certifi_1663615689491/work/certifi
cffi==1.15.1
chardet==4.0.0
constantly==15.1.0
crochet==1.12.0
cryptography==38.0.1
cssselect==1.1.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==2.0.0
django-mptt==0.12.0
djangorestframework==3.12.4
djongo==1.3.6
et-xmlfile==1.1.0
h11==0.13.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
parsel==1.6.0
price-parser==0.3.4
priority==1.3.0
Protego==0.2.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pymongo==3.12.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2022.2.1
PyYAML==6.0
queuelib==1.6.2
requests==2.25.1
Scrapy==2.5.1
scrapy-selenium==0.0.7
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
tablib==3.2.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.8.0
typing_extensions==4.3.0
tzlocal==2.1
urllib3==1.26.12
validators==0.18.2
w3lib==2.0.1
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0